Fixed missing localURLProvider
Since the bug with the strong reference to the FileProviderExtension was fixed, the FileProviderExtension now actually deinitializes itself. Since we had only a weak reference in the FileProviderAdapter on the FileProviderExtension object, we received thereby after some time no more local URLs. This caused all actions that call an importDocument to fail.